Added new Property RssPluginsViewModel and several proxy methods to it
RssPlugins need to be kept simple as to not clutter up the code for a plugin. Plugins should be kept free of code other than the bare essentials for them to function. This means they should not contain UpdateProperty code for UI updates.

As RssFeedsViewModel creates the instance /handles article selection it's perfect for having some proxy properties the view can use