No longer using a flyout for adding rules but the UI controls already there
Realised that copying the WebUI way of doing things doens't make sense. There's no need to add a rule with just a name before adding the other fields.
Since the flyout would duplicate the fields already present it makes more sense to use the form already there.

Since a rule needs to be selected to be deleted and the space for controls is constrained around the selection - added a delete button to the form and removed the button next to the combobox. Opted for a dialog for confirmation rather than the previous flyout.